Disable -Waddress-of-packed-member, for tNDSHeader
The struct doesn't actually need to be packed. But because it is defined as
such, the compiler warns that it can't be cast to a uint32 without
potential misaligned memory accesses. This is not a concern for the address
of the NDS header that's actually used (0x027FFE00), so disable the
warning.

Signed-off-by: Michael "Chishm" Chisholm <chishm@chishm.com>