showCheatFolder: fix message display when menuLevel.selected == -1
The key to this issue is that menuLevel.selected covers the range of
possible entry indexes, but also  includes the extra value of
CHEAT_MENU_FOLDER_UP, which is -1. When that entry is selected,
previously the note/name were still displayed by checking
contents[menuLevel.selected], which indexes contents with -1 and ends up
reading a bogus value (Which crashes NitroHax depending on the loaded
XML and likely other factors).

The simple fix is to check the value is equal to CHEAT_MENU_FOLDER_UP
before attempting to display the name, and if it is, display a hardcoded
string instead.